+++ package.json
@@ -21,6 +21,7 @@
     "framer-motion": "^10.12.7",
     "graphql": "^16.6.0",
     "js-cookie": "^3.0.5",
+    "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-calendar": "^4.2.1",
     "react-datepicker": "^4.11.0",
@@ -31,6 +32,7 @@
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
+    "@types/react-datepicker": "^4.11.2",
     "@types/react-dom": "^18.0.11",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
     "@typescript-eslint/parser": "^5.57.1",
